[Move] dto files to a separate package
